feat: useSuffixIcon
